[SIG-1775] moved formatRoles to separate file
